Move getConvertToItems() to conversion module
