+++ src/components/modules/index.ts
@@ -20,6 +20,7 @@ import UiAPI from './api/ui';
 import BlockSettings from './toolbar/blockSettings';
 import Toolbar from './toolbar/index';
 import InlineToolbar from './toolbar/inline';
+import Conversion from './toolbar/conversion';
 
 /** . */
 import BlockEvents from './blockEvents';
@@ -60,6 +61,7 @@ export default {
   BlockSettings,
   Toolbar,
   InlineToolbar,
+  Conversion,
 
   // Modules
   BlockEvents,
